+++ ICD.Connect.API/ApiConsole.cs
@@ -200,25 +200,31 @@ namespace ICD.Connect.API
 			DateTime installDateTime = ProgramUtils.ProgramInstallDate;
 			string installDate = installDateTime.ToString("f");
 
-			TimeSpan progUptime = ProcessorUtils.GetProgramUptime();
-			string uptime = string.Format("{0} days {1:D2}:{2:D2}:{3:D2}",
+			DateTime? programStartTime = ProcessorUtils.GetProgramStartTime();
+			TimeSpan progUptime = programStartTime.HasValue
+				                      ? IcdEnvironment.GetUtcTime() - programStartTime.Value
+				                      : TimeSpan.Zero;
+			string progUptimeString = string.Format("{0} days {1:D2}:{2:D2}:{3:D2}",
 			                              progUptime.Days,
 			                              progUptime.Hours,
 			                              progUptime.Minutes,
 			                              progUptime.Seconds);
 
-			TimeSpan systemUptime = ProcessorUtils.GetSystemUptime();
-			string lastRestart = string.Format("{0} days {1:D2}:{2:D2}:{3:D2}",
-			                                   systemUptime.Days,
-			                                   systemUptime.Hours,
-			                                   systemUptime.Minutes,
-			                                   systemUptime.Seconds);
+			DateTime? systemStartTime = ProcessorUtils.GetSystemStartTime();
+			TimeSpan systemUptime = systemStartTime.HasValue
+				                        ? IcdEnvironment.GetUtcTime() - systemStartTime.Value
+				                        : TimeSpan.Zero;
+			string systemUptimeString = string.Format("{0} days {1:D2}:{2:D2}:{3:D2}",
+				                                                           systemUptime.Days,
+				                                                           systemUptime.Hours,
+				                                                           systemUptime.Minutes,
+				                                                           systemUptime.Seconds);
 
 			TableBuilder builder = new TableBuilder("Item", "Value");
 
 			builder.AddRow("Install Date", installDate);
-			builder.AddRow("Current Uptime", uptime);
-			builder.AddRow("Time Since Last Restart", lastRestart);
+			builder.AddRow("Current Program Uptime", progUptimeString);
+			builder.AddRow("Current System Uptime", systemUptimeString);
 
 			return builder.ToString();
 		}